Realizada atualização no código para melhor entendimento nos estudos

+++ Program.cs
@@ -1,36 +1,35 @@
 ﻿using ConsoleApp1AplicacaoFinanceira.Entities;
+using var context = new AplicacaoFinanceiraDbContext();
 
-namespace ConsoleApp1AplicacaoFinanceira
+Console.WriteLine("== Salvando AplicacaoFinanceira ==");
+
+
+// Cria um novo objeto e salva na lista
+AplicacaoFinanceira ap = new("CDB", 100, false);
+context.AplicacaoFinanceiraS.Add(ap);
+Console.WriteLine("== Recuperando AplicacaoFinanceira ==");
+
+
+// Seleciona todos os dados da lista
+Console.WriteLine("== SELECT * ==");
+var DadosRecuperados = context.AplicacaoFinanceiraS.ToList();
+DadosRecuperados.ForEach(x => Console.WriteLine(x));
+
+
+// Seleciona um dado específico da lista
+Console.WriteLine("== SELECT WHERE ==");
+var dadoRecuperado = context.AplicacaoFinanceiraS.Find(4);
+Console.WriteLine(dadoRecuperado);
+
+
+// Remove um dado específico da lista
+Console.WriteLine("== REMOVE ==");
+if (dadoRecuperado != null)
 {
-    internal class Program
-    {
-        static void Main(string[] args)
-        {
-            using var context = new AplicacaoFinanceiraDbContext();
-
-            Console.WriteLine("== Salvando AplicacaoFinanceira ==");
-
-            AplicacaoFinanceira ap = new ("Poupança",1, false);
-
-            context.AplicacaoFinanceiraS.Add(ap);
-
-            //context.SaveChanges();
-
-            Console.WriteLine("== Recuperando AplicacaoFinanceira ==");
-
-            Console.WriteLine("== SELECT * ==");
-            var DadosRecuperados =context.AplicacaoFinanceiraS.ToList();
-            DadosRecuperados.ForEach(x => Console.WriteLine(x));
-
-            Console.WriteLine("== SELECT WHERE ==");
-            var dadoRecuperado = context.AplicacaoFinanceiraS.Find(3);
-            Console.WriteLine(dadoRecuperado);
-
-            context.AplicacaoFinanceiraS.Remove(dadoRecuperado);
-
-            Console.WriteLine("== SELECT * ==");
-            var DadosAtualizados = context.AplicacaoFinanceiraS.ToList();
-            DadosRecuperados.ForEach(x => Console.WriteLine(x));
-        }
-    }
+    context.AplicacaoFinanceiraS.Remove(dadoRecuperado);
+    context.SaveChanges();
+}
+else
+{
+    Console.WriteLine("Nada a remover.");
 }
